// pop-web/src/app/shared/openapi/model/family-history.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { User } from './user';
import { CodedConcept } from './coded-concept';


export interface FamilyHistory { 
    /**
     * 
     */
    id: string;
    /**
     * 
     */
    createdAt: string;
    /**
     * 
     */
    updatedAt: string;
    createdBy?: User | null;
    externalSource?: string | null;
    externalSourceId?: string | null;
    /**
     * Indicates the case of the patient who\'s family\'s history is being recorded
     */
    caseId: string;
    /**
     * Clinically-relevant date at which the patient\'s family history was assessed and recorded.
     */
    date: string;
    /**
     * Relationship to the patient
     */
    relationship: CodedConcept;
    /**
     * Whether the family member has a history of cancer
     */
    hadCancer: boolean;
    contributedToDeath?: boolean | null;
    onsetAge?: number | null;
    topography?: CodedConcept | null;
    morphology?: CodedConcept | null;
    /**
     * The user(s) who updated the data since its creation
     */
    updatedBy?: Array<User>;
    /**
     * Human-readable description
     */
    description: string;
}