// pop-web/src/app/shared/openapi/model/family-history-create.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CodedConcept } from './coded-concept';


export interface FamilyHistoryCreate { 
    externalSource?: string | null;
    externalSourceId?: string | null;
    /**
     * Indicates the case of the patient who\'s family\'s history is being recorded
     */
    caseId: string;
    /**
     * Clinically-relevant date at which the patient\'s family history was assessed and recorded.
     */
    date: string;
    /**
     * Relationship to the patient
     */
    relationship: CodedConcept;
    /**
     * Whether the family member has a history of cancer
     */
    hadCancer: boolean;
    contributedToDeath?: boolean | null;
    onsetAge?: number | null;
    topography?: CodedConcept | null;
    morphology?: CodedConcept | null;
}